% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarized_experiment.R
\name{attach_coldata}
\alias{attach_coldata}
\title{Attach columns to existing data.frame and checks whether row and column names
are consistent.}
\usage{
attach_coldata(se_coldata, extra_coldata)
}
\arguments{
\item{se_coldata}{the original data.frame to extend}

\item{extra_coldata}{a data.frame with the additional columns to add. Row
and column names must be consistent with the original data.frame, i.e., the
extra_coldata must include all row names and none colnames of the original
data.frame}
}
\value{
a data.frame with the extra columns attached
}
\description{
Attach columns to existing data.frame and checks whether row and column names
are consistent.
}
\examples{
\dontrun{
extra_cols <-
  data.frame(condition = rep_len("A", rownames(colData(se))),
             row.names = rownames(colData(se)))
colData(se) <- attach_coldata(colData(se), extra_cols)
}
}
