% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{cigar_dist}
\alias{cigar_dist}
\title{Compute base pairs distance from CIGAR string}
\usage{
cigar_dist(cigar_str)
}
\arguments{
\item{cigar_str}{a CIGAR string}
}
\description{
Compute base pairs distance from CIGAR string
}
