% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{combine_circrna_methods}
\alias{combine_circrna_methods}
\title{Combine outputs of different circRNA detection tools}
\usage{
combine_circrna_methods(
  file_names,
  filter = TRUE,
  min_reads = 2,
  min_methods = 2
)
}
\arguments{
\item{file_names}{a names list of the file names to be combined}

\item{filter}{if TRUE performs a filtering on the circRNAs included in the
results: only circRNAs with >= \code{min_reads} and detected by >=
\code{min_methods} will be included}

\item{min_reads}{the minimum read count a circRNA must have to be included in
the results when the \code{filter} option is enabled}

\item{min_methods}{the minimum number of methods a circRNA must be detected
by to be included in the results when the \code{filter} option is enabled}
}
\description{
Combine outputs of different circRNA detection tools
}
