% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ccp2_runs.R
\name{compute_lin_bks_counts}
\alias{compute_lin_bks_counts}
\title{Count the reads linearly spliced on the backsplice junctions}
\usage{
compute_lin_bks_counts(
  files,
  circ_ids,
  is_stranded = TRUE,
  is_paired_end = TRUE,
  cpus = 1
)
}
\arguments{
\item{files}{the CirComPara2 project directories}

\item{circ_ids}{the list of backsplices in the form of chr:start-end[:strand]}

\item{is_stranded}{should the strand be considered? (default: TRUE)}

\item{is_paired_end}{were the reads mapped as paired-end? (default: TRUE)}

\item{cpus}{number of CPUs for parallel execution (default: 1)}
}
\value{
a matrix with circRNAs as rows and samples as columns and counts in 
the cells
}
\description{
Count the reads linearly spliced on the backsplice junctions
}
