% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{dcc_fix_strand}
\alias{dcc_fix_strand}
\title{Add strand to DCC output file}
\usage{
dcc_fix_strand(counts, coordinates, output = "")
}
\arguments{
\item{counts}{the CircRNACounts file resulting from DCC}

\item{coordinates}{the CircCoordinates file resulting from DCC}

\item{output}{the output file name}
}
\value{
the CircRNACounts table file with a "Strand" column for the strand
}
\description{
Attach strand column to DCC's CircRNACount file. When the CircRNACounts DCC
output file does not report the circRNA strand, the strand information can be
retrieved from the CircCoordinates file.
}
\author{
Enrico Gaffo
}
