% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCLP_test.R
\name{DCLP_test}
\alias{DCLP_test}
\title{Differential circular-to-linear proportion tests}
\usage{
DCLP_test(circ, lin, groups, design = "~ 1", ...)
}
\arguments{
\item{circ}{matrix with backsplice junction read counts}

\item{lin}{matrix with the counts of the linearly spliced reads on backsplice
junction}

\item{groups}{data.frame with samples covariates}

\item{design}{character with the model formula}

\item{...}{additional parameters passed to BiocParallel::bplapply()}
}
\value{
a named vector of P-values
}
\description{
Differential circular-to-linear proportion tests run in parallel
}
\examples{
\dontrun{
# set BiocParallel parameters for parallel execution
biocParams <- BiocParallel::SnowParam(workers = 8, "SOCK", 
                                      progressbar = FALSE)

# consider a design formula including batch effects
Groups <- data.frame(samples = c("sample1", "sample2", "sample3", "sample4"),
                     batch = c("batch1", "batch1", "batch2", "batch2"),
                     condition = c("WT", "Treated", "WT", "Treated"))
design <- "~ batch + condition"

# get inputs: BJR read counts and linearly spliced reads on the BJ
prjs <- c("/home/user/circompara2_batch1",
          "/home/user/circompara2_batch2")
           
## merge the projects           
combs <- combine_ccp2_runs(prjs)

## circRNA BJR matrix 
Circ <- as.matrix(data.frame(combs$circ_read_count_mt,
                             row.names = "circ_id")) 
Linear <- as.matrix(data.frame(combs$lin_read_count_mt, 
                               row.names = "circ_id"))

# compute differential CLPs
DCLP_pvals <- 
  DCLP_test(circ = Circ, 
           lin = Linear, 
           groups = Groups, 
           design = design, 
           BPPARAM = biocParams)
            
# compute adjusted P-values
DCLP_padj <- p.adjust(p = DCLP_pvals, method = "BH")
}
}
