% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{filter_and_cast_circexp}
\alias{filter_and_cast_circexp}
\title{Filter the CirComPara2 output and reshape it into a rectangular matrix format}
\usage{
filter_and_cast_circexp(
  input_file,
  min_reads = 2,
  min_methods = 2,
  output_file = "ce2.csv"
)
}
\arguments{
\item{input_file}{CircRNA expression table file in 'long' format.}

\item{min_reads}{The minimum detection read threshold for a circRNA
in at least one sample}

\item{min_methods}{Keep circRNAs commonly detected by >= min_methods circRNA
detection methods in at least one sample}

\item{output_file}{output directory}
}
\description{
Filter CirComPara2 predictions and reshape into matrix format after filtering
by minimum read count and number of detecting methods.
Stranded input is automatically detected
}
