% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{filter_findcirc_res}
\alias{filter_findcirc_res}
\title{Filter Findcirc output}
\usage{
filter_findcirc_res(
  input,
  output = "filteredFindcirc.bed",
  minqual = 40,
  flags = "CIRCULAR"
)
}
\arguments{
\item{input}{Findcirc result table file (circ_candidates.bed)}

\item{output}{The output file}

\item{minqual}{Minimum quality of backsplice end mapping}

\item{filter_tags}{Comma separated list of tags that have to be present in
rows of the Find_circ BED file}
}
\value{
a table file
}
\description{
Filter Findcirc output to keep good circRNAs
}
