% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_quasibinomial.R
\name{fit_quasibinomial}
\alias{fit_quasibinomial}
\title{Fit a GLM of the quasibinomial family}
\usage{
fit_quasibinomial(id, x, mod)
}
\arguments{
\item{id}{an identifier to select the matrix row}

\item{x}{a numeric matrix}

\item{mod}{a matrix representing the model to fit. For instance, as returned
by the \code{\link[=model.matrix]{model.matrix()}} function}
}
\value{
an object of class inheriting from "glm", as the output of the
\code{\link[=glm.fit]{glm.fit()}} function
}
\description{
Auxiliary function to fit a Generalized Linear Model of the quasibinomial
family to a row of a matrix
}
\author{
Enrico Gaffo
}
