% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{format_ce2}
\alias{format_ce2}
\title{Format CIRCexplorer2 output into simple "circ_id, read_count" table}
\usage{
format_ce2(input_file)
}
\value{
a data.table with two columns: circ_id and read_count
}
\description{
Format CIRCexplorer2 output into simple "circ_id, read_count" table
}
