% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{format_cf}
\alias{format_cf}
\title{Format circRNA_finder output into simple "circ_id, read_count" table}
\usage{
format_cf(input_file)
}
\value{
a data.table with two columns: circ_id and read_count
}
\description{
Format circRNA_finder output into simple "circ_id, read_count" table
}
