% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{format_ciri}
\alias{format_ciri}
\title{Format CIRI2 output into simple "circ_id, read_count" table}
\usage{
format_ciri(input_file)
}
\value{
a data.table with two columns: circ_id and read_count
}
\description{
Format CIRI2 output into simple "circ_id, read_count" table
}
