% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{format_dcc}
\alias{format_dcc}
\title{Format DCC output into simple "circ_id, read_count" table}
\usage{
format_dcc(input_file)
}
\value{
a data.table with two columns: circ_id and read_count
}
\description{
Format DCC output into simple "circ_id, read_count" table
}
