% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{format_fc}
\alias{format_fc}
\title{Format Find_circ output into simple "circ_id, read_count" table}
\usage{
format_fc(input_file)
}
\value{
a data.table with two columns: circ_id and read_count
}
\description{
Format Find_circ output into simple "circ_id, read_count" table
}
