% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{gene_annotation}
\alias{gene_annotation}
\title{Title}
\usage{
gene_annotation(
  combined_circrnas = "combined_circrnas.gtf.gz",
  cluster_dist = 5000L,
  outdir = "./"
)
}
\arguments{
\item{cluster_dist}{The maximum distance allowed to cluster intergenic circRNAs}

\item{outdir}{Output directory were circ_to_genes.tsv and gene_to_circ.tsv will be saved}
}
\description{
Title
}
