% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{get.avg.read.len}
\alias{get.avg.read.len}
\title{Compute average length}
\usage{
get.avg.read.len(fastqc_data.txt.file)
}
\arguments{
\item{fastqc_data.txt.file}{}
}
\description{
Parse a fastqc_data.txt file and computes the average read length
TODO: improve by parsing "fastqc_data.txt" and get the mode instead of
average read length
}
