% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{get.circ.cluster}
\alias{get.circ.cluster}
\title{Cluster intergenic circrnas}
\usage{
get.circ.cluster(intergenic.circs.bed, max.dist = 5000L)
}
\arguments{
\item{intergenic.circs.bed}{a data.table in BED format}

\item{max.dist}{The maximum distance allowed between circRNAs of a same
cluster}
}
\value{
a data.table
}
\description{
Assign a common label to neighboring intergenic circRNAs
}
