% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{get_bks_reads_ce2_bwa}
\alias{get_bks_reads_ce2_bwa}
\title{Obtain the backspliced reads for the circRNAs predicted by CIRCexplorer2 on
BWA mappings}
\usage{
get_bks_reads_ce2_bwa(
  chimreads,
  circrnas,
  output = "circular.reads.bed.gz",
  range = 10,
  discriminate_mates = FALSE,
  stranded = FALSE
)
}
\arguments{
\item{chimreads}{sample_bwa.sam.gz BED file transformed by
get_ce2_bwa_circ_reads.py. The input file bearing the backspliced reads}

\item{circrnas}{circularRNA_known.txt as output by CIRCexplorer2 annotate,
or back_spliced_junction.bed as output by CIRCexplorer2 parse}

\item{output}{The circRNA read IDs for each circRNA in compressed BED}

\item{range}{Number of basepairs tolerated in realigning circRNAs from
CIRCexplorer2 annotate}

\item{discriminate_mates}{By default, reads will be collapsed by read ID and
position. Enable this option to discriminate paired-end read mates by
appending \1 or \2 to read IDs.}

\item{stranded}{Do not consider strand of alignments. The reported strand
will be copied from CIRCexplorer2 annotate results}
}
\description{
Obtain the backspliced reads for the circRNAs predicted by CIRCexplorer2 on
BWA mappings
}
