% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{get_bks_reads_ce2_segemehl}
\alias{get_bks_reads_ce2_segemehl}
\title{Obtain the backspliced reads for the circRNAs predicted by CIRCexplorer2 on
Segemehl mappings}
\usage{
get_bks_reads_ce2_segemehl(
  chimreads,
  circrnas,
  output = "circular.reads.bed.gz",
  range = 10,
  discriminate_mates = FALSE
)
}
\arguments{
\item{chimreads}{sample.sngl.bed plus sample.trns.txt merged bks read as
from filter_testrealign.R}

\item{circrnas}{circularRNA_known.txt as output by CIRCexplorer2 annotate,
or back_spliced_junction.bed as output by CIRCexplorer2 parse}

\item{output}{The circRNA read IDs for each circRNA in compressed BED}

\item{range}{Number of basepairs tolerated in realigning circRNAs from
CIRCexplorer2 annotate}

\item{discriminate_mates}{By default, reads will be collapsed by read ID and
position. Enable this option to discriminate paired-end read mates by
appending \1 or \2 to read IDs.}
}
\description{
Obtain the backspliced reads for the circRNAs predicted by CIRCexplorer2 on
Segemehl mappings
}
