% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{get_bks_reads_circrnaFinder}
\alias{get_bks_reads_circrnaFinder}
\title{Obtain the backspliced reads for the circRNAs predicted by circRNA_finder}
\usage{
get_bks_reads_circrnaFinder(
  chimreads,
  circrnas,
  output = "circular.reads.bed.gz"
)
}
\arguments{
\item{chimreads}{Chimeric.out.junction BED filtered and formatted by
cf_filterChimout.awk}

\item{circrnas}{circularRNA_known.txt as output by CIRCexplorer2 annotate,
or back_spliced_junction.bed as output by CIRCexplorer2 parse}

\item{output}{The circRNA read IDs for each circRNA in compressed BED}
}
\description{
Obtain the backspliced reads for the circRNAs predicted by circRNA_finder
}
