% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{get_bks_reads_dcc}
\alias{get_bks_reads_dcc}
\title{Obtain the backspliced reads for the circRNAs predicted by DCC}
\usage{
get_bks_reads_dcc(
  chimreads,
  circrnas,
  output = "circular.reads.bed.gz",
  tolerance = 5,
  stranded = FALSE
)
}
\arguments{
\item{chimreads}{Chimeric.out.junction BED transformed by
chimout_junc_to_bed.py}

\item{circrnas}{strandedCircRNACount as output by DCC}

\item{output}{The circRNA read IDs for each circRNA in compressed BED}

\item{tolerance}{Number of base pairs tolerated in realigning circRNAs}

\item{stranded}{Set TRUE if stranded library}
}
\description{
Obtain the backspliced reads for the circRNAs predicted by DCC
}
