% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{get_bks_reads_findcirc}
\alias{get_bks_reads_findcirc}
\title{Obtain the backspliced reads for the circRNAs predicted by Find_circ}
\usage{
get_bks_reads_findcirc(
  sites_reads = "sites.reads",
  circ_candidates = "circ_candidates.bed",
  reads_bed = "circular.reads.bed.gz",
  read_list = "bks.reads"
)
}
\arguments{
\item{sites_reads}{Read file (in FASTA format) as the ouput by -R option of
find_circ.py}

\item{circ_candidates}{BED file with backsplices, as find_circ.py main output}

\item{reads_bed}{Output filename of BED file with read IDs for each backsplice}

\item{read_list}{Output filename of all read ID list}
}
\description{
Obtain the backspliced reads for the circRNAs predicted by Find_circ
}
