% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{get_bks_reads_per_method}
\alias{get_bks_reads_per_method}
\title{Get the total backspliced read number for each method}
\usage{
get_bks_reads_per_method(bks_read_file)
}
\arguments{
\item{bks_read_file}{the file collecting the backspliced reads, i.e.,
the bks.counts.collected_reads.csv file}
}
\value{
a data.table with sample_id, circ_method and n_reads columns
}
\description{
Compute the histogram of circRNA expression detected by each method,
per each sample
}
\author{
Enrico Gaffo
}
