% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ccp2_runs.R
\name{get_circrna_host_genes}
\alias{get_circrna_host_genes}
\title{Find circRNA host-gene(s)}
\usage{
get_circrna_host_genes(circ_ids, gtf_file)
}
\arguments{
\item{circ_ids}{a list of circRNA identifiers in the form of
chr:start-end[:strand]}

\item{gtf_file}{an Ensemble GTF gene annotation file}
}
\value{
a data.table of the genes overlapping each circRNA, one row for each
overlap.
}
\description{
Find circRNA host-gene(s)
}
\examples{
\dontrun{
get_circrna_host_genes("10:1000676-1000868:+",
                       "Homo_sapiens.GRCh38.108.gtf")}
}
