% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{get_circrnas_per_method}
\alias{get_circrnas_per_method}
\title{Get the number of circRNAs identified by each method
in each sample}
\usage{
get_circrnas_per_method(circrnas_gtf, minreads = 2)
}
\arguments{
\item{circrnas_gtf}{the circrnas.gtf generated by CirComPara2}

\item{minreads}{do not count circRNAs with less than \emph{minreads}}
}
\value{
a data.table with three columns: sample_id, Method, and n_circrnas
}
\description{
A function to count the circRNAs detected by each of the
circRNA methods applied
}
\author{
Enrico Gaffo
}
