% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{get_gene_expression_files}
\alias{get_gene_expression_files}
\title{Title}
\usage{
get_gene_expression_files(
  transcripts.gtf.files,
  gene_raw_counts_list,
  gene.xpr.file,
  trx_raw_counts_list,
  results.dir = "./"
)
}
\arguments{
\item{transcripts.gtf.files}{The transcripts.gtf files as output by Stringtie}

\item{results.dir}{The output directory path}
}
\description{
Title
}
