% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{get_linear_bks_reads}
\alias{get_linear_bks_reads}
\title{Get the circRNA parent-gene linear expression}
\usage{
get_linear_bks_reads(
  lin_reads,
  output_path = ".",
  output_filename = "ccp_bks_linexp.csv"
)
}
\arguments{
\item{lin_reads}{a text file listing the linear expression table files, or
the linear expression table itself. It accepts both CirComPara and DCC
tables.}

\item{output_path}{the directory where to save the results}

\item{output_filename}{the name of the output file}
}
\description{
Process the circRNA parent-gene linear expression files and merge them into a
matrix
}
\author{
Enrico Gaffo
}
