% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{get_stringtie_rawcounts}
\alias{get_stringtie_rawcounts}
\title{Title}
\usage{
get_stringtie_rawcounts(transcripts_gtf, fastqc_data = 100, outprefix = "")
}
\arguments{
\item{transcripts_gtf}{The _transcripts.gtf file output by StringTie}

\item{fastqc_data}{The fastqc_data.txt output by FastQC, or a comma
separated list of fastqc_data.txt files if paired-end reads. If a number is
given then it will be interpreted as the (already computed) average read
length}

\item{outprefix}{The prefix of output file names}
}
\description{
Computes the raw read counts for both transcripts and genes like the
prepDE.py script from StringTie. It will output two files: the transcripts'
and the genes' raw read counts.
}
