% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{guess_input_method}
\alias{guess_input_method}
\title{Guess the circRNA detection method from its output file.}
\usage{
guess_input_method(input_file)
}
\arguments{
\item{input_file}{the output of any circRNA detection method among
CIRCexplorer2, circRNA_finder, CIRI2, DCC, Find_circ, and Segemehl}
}
\value{
a string with the name of the guessed method (abbreviations are
used). NA if the guess fails.
}
\description{
Guess the circRNA detection method from its output file.
}
