% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ccp2_runs.R
\name{guess_strand}
\alias{guess_strand}
\title{Try to resolve strandedness ambiguities of the detected circRNAs}
\usage{
guess_strand(strand_pattern, circ_methods = NULL)
}
\arguments{
\item{strand_pattern}{a character in the form of 
"strand_reads_nMethods|strand_reads_nMethods". E.g. "+_30_6|-_26_1"}

\item{circ_methods}{(optional) the methods corresponding to the strand 
pattern, separated by an "@" character. It helps to solve some ambiguities.
E.g: "CIRCexplorer2_bwa|CIRCexplorer2_star|CIRCexplorer2_tophat@dcc"}
}
\value{
"+", "-", or "." when the ambiguity cannot be resolved
}
\description{
Try to resolve strandedness ambiguities of the detected circRNAs
}
\examples{
guess_strand("+_30_6|-_26_1") # "+"
guess_strand("+_30_6|-_36_1") # "-"
guess_strand("+_30_6|-_30_1") # "+"
guess_strand("+_30_1|-_30_1") # "."
guess_strand("+_30_6") # "+"
}
