% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{melt_dcc_linexp_tab}
\alias{melt_dcc_linexp_tab}
\title{Reshape a DCC linear expression matrix into long format}
\usage{
melt_dcc_linexp_tab(x)
}
\arguments{
\item{x}{DCC linear expression matrix (rows are circRNAs, columns are
samples)}
}
\value{
the linear expression in a long table format
(three columns: circ_id, sample_id, lin.reads)
}
\description{
Auxiliary function to transform a DCC linear expression matrix into
long table format
}
\author{
Enrico Gaffo
}
