% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ccp2_runs.R
\name{merge_ccp_counts}
\alias{merge_ccp_counts}
\title{Get circRNA BJR count expression matrix}
\usage{
merge_ccp_counts(files, groups = NULL)
}
\arguments{
\item{files}{the path of a CirComPara2 run, or a text file listing the paths
of multiple CirComPara2 run to merge.}

\item{is_list_file}{set TRUE if "files" is a text file listing the projects
to merge}
}
\value{
A \code{data.table} with the following columns:
\itemize{
\item \code{sample_id}: Sample identifier.
\item \code{chr}, \code{start}, \code{end}, \code{strand}: Genomic coordinates and strand.
\item \code{read.count}: Backsplice junction read count (BJR).
\item \code{n_methods}: Number of circRNA-detection tools that commonly identified the circRNA.
\item \code{circ_methods}: Names of the circRNA-detection tools that commonly identified the circRNA,
in a bar-separated list, alphanumerically sorted.
Example: \code{"CIRCexplorer2_bwa|ciri_out|dcc"}.
\item \code{circ_id}: circRNA identifier composed as \code{chr:start-end:strand}.
\item \code{strand_pattern}: A string representing the number of reads and methods according to their strand(s).
See also \link[ccp2tools:guess_strand]{guess_strand}.
}
}
\description{
Compute the backsplice junction read (BJR) counts according to the
CirComPara2's framework. This function collect the results of one or more
CirComPara2 runs and merges the results.
}
