% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{merge_ciri}
\alias{merge_ciri}
\title{Combine CIRI2 outputs}
\usage{
merge_ciri(labels, inputs, output = "")
}
\arguments{
\item{labels}{a comma-separated string listing the sample names of the files
to be merged}

\item{inputs}{a comma-separated string listing the files to be merged. Input
files must be _ciri.out as resulting from CIRI2}

\item{output}{the file name of the combined table}
}
\value{
nothing. Write the combined table into a file in the disk.
}
\description{
Combine CIRI2 outputs of multiple samples into a single table
}
\author{
Enrico Gaffo
}
