% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{merge_findcirc}
\alias{merge_findcirc}
\title{Combine Findcirc outputs}
\usage{
merge_findcirc(labels, inputs, output = "")
}
\arguments{
\item{labels}{a comma-separated string listing the sample names of the files
to be merged}

\item{inputs}{a comma-separated string listing the files to be merged. Input
files must be circ_candidates.bed as resulting from find_circ.py (Findcirc)}

\item{output}{the file name of the combined table}
}
\value{
nothing. Write the combined table into a file in the disk.
}
\description{
Combine Findcirc outputs of multiple samples into a single table
}
\author{
Enrico Gaffo
}
