% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ccp2_runs.R
\name{merge_lin_bks_counts}
\alias{merge_lin_bks_counts}
\title{Merge the counts of the reads linearly spliced on the backsplice junctions}
\usage{
merge_lin_bks_counts(files)
}
\arguments{
\item{files}{the paths of CirComPara2 runs to merge.}
}
\description{
Merge the counts of the reads linearly spliced on the backsplice junctions
}
