% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ccp2_runs.R
\name{merge_lin_counts}
\alias{merge_lin_counts}
\title{Merge multiple CirComPara2 runs to get linear transcript/gene expression}
\usage{
merge_lin_counts(prj_paths, groups = NULL, ...)
}
\arguments{
\item{...}{additional parameters used in tximport::tximport()}
}
\value{
A named list with two elements:
\itemize{
\item \code{txi}: A \link[tximport:tximport]{tximport::tximport} object at transcript level.
\item \code{tx2gene}: A table of transcript/gene IDs used in the \code{txi} object.
}
}
\description{
Merge multiple CirComPara2 runs to get linear transcript/gene expression
}
\examples{
\dontrun{
prjs <- c("/home/user/CirComPara2/batch1",
          "/home/user/CirComPara2/batch2")

## get the linear transcripts' expression
trx_ls <- merge_lin_counts(prjs)

trx <- trx_ls$txi

## get the linear gene expression#'
gnx <- tximport::summarizeToGene(trx_ls$txi, trx_ls$tx2gene)
}
}
