% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ccp2_runs.R
\name{merge_read_stats}
\alias{merge_read_stats}
\title{Merge read statistics (preprocessing and alignments) of multiple CirComPara2
analyses}
\usage{
merge_read_stats(prj_dirs, groups = NULL)
}
\arguments{
\item{prj_dirs}{a character vector of the CirComPara2 runs' directory paths
to be merged.
\code{read_statistics/read_stats_collect/processing_and_mapped_read_counts.csv}
files will be retrieved from each directory of the \code{prj_dirs} list.}
}
\value{
a \code{data.table} of read statistics, one row per sample
}
\description{
Merge read statistics (preprocessing and alignments) of multiple CirComPara2
analyses
}
\examples{
\dontrun{
prj_dirs <- c("/ccp2_runs/SRR3162129/",
              "/ccp2_runs/SRR3162130")
read_stats <- merge_read_stats(prj_dirs)
}

}
