% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{merge_segemehl}
\alias{merge_segemehl}
\title{Combine Segemehl outputs}
\usage{
merge_segemehl(labels, inputs, output = "", fixstart = T)
}
\arguments{
\item{labels}{a comma-separated string listing the sample names of the files
to be merged}

\item{inputs}{a comma-separated string listing the files to be merged. Input
files must be splicesites.bed as resulting from Segemehl}

\item{output}{the file name of the combined table}

\item{fixstart}{if TRUE (default), will subtract 1 to the backsplice start
position}
}
\value{
nothing. Write the combined table into a file in the disk.
}
\description{
Combine Segemehl outputs of multiple samples into a single table
}
\author{
Enrico Gaffo
}
