% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{paste.bar}
\alias{paste.bar}
\title{Collapse a vector of strings into a string with elements separated by bar
("|") characters}
\usage{
paste.bar(x)
}
\arguments{
\item{x}{a vector of characters}
}
\description{
Collapse a vector of strings into a string with elements separated by bar
("|") characters
}
