% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{read_circrna_input}
\alias{read_circrna_input}
\title{Reads the input file and convert to circ_id, read_count table format}
\usage{
read_circrna_input(input_file)
}
\value{
a "circ_id, read_count" data.table with BED circRNA coordinates
}
\description{
Reads the input file and convert to circ_id, read_count table format
}
