% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_batch_effect_binomial.R
\name{remove_batch_effect_binomial}
\alias{remove_batch_effect_binomial}
\title{Remove Batch Effects by fitting a Binomial GLM}
\usage{
remove_batch_effect_binomial(
  x,
  batch,
  covariates = NULL,
  design = matrix(1, ncol(x), 1),
  ...
)
}
\arguments{
\item{x}{}

\item{batch}{}

\item{covariates}{}

\item{design}{}

\item{...}{}
}
\description{
A function to remove batch effects from a matrix of proportion
values. Like the homonymous function from the limma package but for values in
the range [0..1]. Useful to correct batch effects from CLP matrices.
}
\author{
Enrico Gaffo
}
