% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_batch_effect_quasibin.R
\name{remove_batch_effect_quasibin}
\alias{remove_batch_effect_quasibin}
\title{Remove Batch Effects by fitting a Qquasi-binomial GLM}
\usage{
remove_batch_effect_quasibin(
  x,
  batch,
  covariates = NULL,
  design = matrix(1, ncol(x), 1),
  ...
)
}
\arguments{
\item{x}{numeric matrix containing proportion (i.e., real numbers between 0
and 1) values for a series of samples. Rows correspond to probes and
columns to samples.}

\item{batch}{factor or vector indicating batches.}

\item{covariates}{matrix or vector of numeric covariates to be adjusted for.}

\item{design}{model design matrix}

\item{...}{not used}
}
\value{
A numeric matrix of proportion values with batch and covariate
effects removed.
}
\description{
A function to remove batch effects from a matrix of proportion values. Like
the homonymous function from the limma package but for values between 0 and 1.
Useful to correct batch effects from CLP matrices.
}
\seealso{
limma::removeBatchEffect()
}
\author{
Enrico Gaffo
}
