% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{simplify.gene.region.list}
\alias{simplify.gene.region.list}
\title{Simplify gene region annotation character vectors}
\usage{
simplify.gene.region.list(x)
}
\arguments{
\item{x}{the vector of labels}
}
\description{
Simplify gene region annotation character vectors
}
