% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{simplify.gene.region}
\alias{simplify.gene.region}
\title{Simplify gene region annotation character labels}
\usage{
simplify.gene.region(x)
}
\arguments{
\item{x}{the label to be simplified}
}
\description{
Simplify gene region annotation character labels
}
