% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccp2tools.R
\name{switch_strand}
\alias{switch_strand}
\title{Switch between forward and reverse strands}
\usage{
switch_strand(s)
}
\arguments{
\item{s}{the strand to be inverted. Must be "+" or "-"}
}
\value{
the opposite of the input strand
}
\description{
A function to invert the strand sign
}
