% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCLP_test.R
\name{test_betabin}
\alias{test_betabin}
\title{Test circular-to-linear proportion variation}
\usage{
test_betabin(id, circ, lin, group, full, red)
}
\arguments{
\item{id}{circRNA id to test}

\item{circ}{matrix with backsplice junction read counts}

\item{lin}{matrix with the counts of the linearly spliced reads on backsplice
junction}

\item{group}{data.frame with samples covariates}

\item{full}{character with the full model formula}

\item{red}{character with the reduced model formula}
}
\value{
numeric P-value
}
\description{
Fit and compare Beta binomial models
}
