% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarized_experiment.R
\name{to_summarized_experiment}
\alias{to_summarized_experiment}
\title{Convert the result of \link{combine_ccp2_runs} into a SummarizedExperiment object}
\usage{
to_summarized_experiment(
  combined_prjs,
  assay = c("circrna", "gene", "trx"),
  coldata = NULL
)
}
\arguments{
\item{combined_prjs}{A named list, as returned by the \link{combine_ccp2_runs} function.
Must include the following entries:
\itemize{
\item \code{circ_read_count_mt}
\item \code{lin_read_count_mt}
\item \code{circ_gene_anno}
\item \code{lin_xpr}
\item \code{read_stats}
\item \code{gene_anno}
}}

\item{assay}{select which RNA assay to extract among "circrna", "gene", or
"trx" for circrna-, gene-, or transcript-level expression analysis.}

\item{coldata}{a data.frame with additional sample data to include into the
\link{SummarizedExperiment}}
}
\value{
a \link{RangedSummarizedExperiment} object
}
\description{
Convert the result of \link{combine_ccp2_runs} into a SummarizedExperiment object
}
\examples{
\dontrun{

library(ccp2tools)

prjs <- c("/home/user/ccp2_nf/sample1",
          "/home/user/ccp2_nf/sample2")

combined_prjs <- combine_ccp2_runs(prjs)

circrna_se <- to_summarized_experiment(combined_prjs = combined_prjs)

gene_se <- to_summarized_experiment(combined_prjs = combined_prjs,
                                    assay = "gene")

trx_se <- to_summarized_experiment(combined_prjs = combined_prjs,
                                    assay = "trx")
}

}
