% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ccp2_runs.R
\name{merge_lin_counts}
\alias{merge_lin_counts}
\title{Merge multiple CirComPara2 runs to get linear transcript/gene expression}
\usage{
merge_lin_counts(prj_paths, ...)
}
\arguments{
\item{prj_paths}{}

\item{...}{additional parameters used in tximport::tximport()}
}
\value{
a named list of two elements: 
 1) txi: a tximport object at transcript level
 2) tx2gene: the transcript/gene ids table used in the txi object
}
\description{
Merge multiple CirComPara2 runs to get linear transcript/gene expression
}
\examples{
\dontrun{
prjs <- c("/home/user/CirComPara2/batch1",
          "/home/user/CirComPara2/batch2")

## get the linear transcripts' expression
trx_ls <- merge_lin_counts(prjs)

trx <- trx_ls$txi

## get the linear gene expression#' 
gnx <- tximport::summarizeToGene(trx_ls$txi, trx_ls$tx2gene)
}
}
